requirements.txt
gensim
tqdm
pandas
numpy
scipy
networkx
scikit_learn
torchvision
python-igraph